Enabled energy dispersal
Qam-modulator now randomises each data byte using pseudo-random generator.

Added comments to the codec and qam modulator further explaining the functions operation.

+++ src/codec/fft/ofdmfft.cpp
@@ -19,7 +19,7 @@
 * @return 0 on success, else error number
 *
 */
-int ofdmFFT::Configure(uint16_t nPoints, int type, uint32_t pilotStep)
+int ofdmFFT::Configure(size_t nPoints, int type, size_t pilotStep)
 {   
     // If object has been configured before
     if(m_configured)
